% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble.R
\name{scramble}
\alias{scramble}
\title{Scramble}
\usage{
scramble(text, n)
}
\arguments{
\item{text}{character vector}

\item{n}{numeric, number of permutations for each string}
}
\value{
A list of length equal to length of \code{text}, containing character vectors of random permutations each of length \code{n}.
}
\description{
This function provides random permutations of character strings
}
\examples{
scramble(c('vanderbilt biostatistics', 'lightning talks'), 3)

}
