% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scramble_data.R
\name{scramble_data}
\alias{scramble_data}
\title{scramble_data}
\usage{
scramble_data(dat, keep_col_names = TRUE, keep_row_names = TRUE)
}
\arguments{
\item{dat}{A data frame or matrix.}

\item{keep_col_names}{\code{TRUE} or \code{FALSE}, indicating whether to retain
the input data's column names in the scrambled data. (If \code{FALSE}, the
scrambled data's column names are \code{V1}, \code{V2}, etc.)}

\item{keep_row_names}{\code{TRUE} or \code{FALSE}, indicating whether to retain
the input data's row names in the scrambled data. (If \code{FALSE}, the
scrambled data's row names are \code{1}, \code{2}, etc.)}
}
\value{
A data frame or matrix, depending on \code{dat}.
}
\description{
Scrambles the data in a data frame or matrix while retaining the type and
basic properties (approximate min/max for numeric data, approximate
length for character data, approximate \code{NA} proportion, etc.) of the
input. Principled scrambling approaches are defined for logical, integer,
double, and character data. See \code{?scramble_vec} for info on the specific
approach taken for each data type.
}
\examples{
set.seed(248)
scramble_data(mtcars)
scramble_data(USPersonalExpenditure, FALSE, FALSE)

}
